+++ app/lib/src/features/dashboard/widgets/pumps/pump_pressure_rate.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:irrigazione_iot/src/features/dashboard/widgets/dashboard_child_item_details_row.dart';
+import 'package:irrigazione_iot/src/features/dashboard/widgets/shared/dashboard_child_item_details_row.dart';
 import 'package:irrigazione_iot/src/features/dashboard/widgets/shared/common_pressure_rate_trailing_text.dart';
 import 'package:irrigazione_iot/src/features/pumps/data/pump_statistic_repository.dart';
