+++ app/lib/src/features/dashboard/widgets/shared/dashboard_child_item_details_row.dart
